KERNEL HOTPLUGS: added control for Suspend or Always active! read more.
Some of my users report that they cant wake the device in random cases.
All started when i have fixed the suspend mode as it's was not working
for long time.
But in the same time a lot of people reported all good and stable!
including me when using the hotplugs in suspend mode. no SODS!

So by default it's should be active, OK! let it be active.
I have added STweaks control for it. with info and warnings.
Also using hotplug in active mode,
will allow faster screen wake when call is coming in.
Need to check battery but i dont think it's really major saver...

* Additionally i have remove the cpu boost when suspend is canceled
it's seems that in some condition we get NULL error on _cpu_up
so use the boost only after real suspend time...